add `useIsTeamAdmin` hook
